// types/OmnivoreSchema.ts
export type Maybe<T> = T | null
export type Exact<T extends Record<string, unknown>> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
{ [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
{ [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  Date: any
}

export enum SortOrder {
  Ascending = 'ASCENDING',
  Descending = 'DESCENDING'
}

export enum ReactionType {
  Like = 'LIKE',
  Heart = 'HEART',
  Smile = 'SMILE',
  Hushed = 'HUSHED',
  Crying = 'CRYING',
  Pout = 'POUT'
}

export enum SortBy {
  UpdatedTime = 'UPDATED_TIME',
  Score = 'SCORE',
  SavedAt = 'SAVED_AT',
  PublishedAt = 'PUBLISHED_AT'
}

export enum ContentReader {
  Web = 'WEB',
  Pdf = 'PDF',
  Epub = 'EPUB'
}

export interface SortParams {
  order?: Maybe<SortOrder>
  by: SortBy
}

export interface PageInfo {
  __typename?: 'PageInfo'
  hasNextPage: Scalars['Boolean']
  hasPreviousPage: Scalars['Boolean']
  startCursor?: Maybe<Scalars['String']>
  endCursor?: Maybe<Scalars['String']>
  totalCount?: Maybe<Scalars['Int']>
}

export interface ArticleEdge {
  __typename?: 'ArticleEdge'
  cursor: Scalars['String']
  node: Article
}

export interface FeedArticleEdge {
  __typename?: 'FeedArticleEdge'
  cursor: Scalars['String']
  node: FeedArticle
}

export interface ArticlesSuccess {
  __typename?: 'ArticlesSuccess'
  edges: ArticleEdge[]
  pageInfo: PageInfo
}

export interface User {
  __typename?: 'User'
  id: Scalars['ID']
  name: Scalars['String']
  isFullUser?: Maybe<Scalars['Boolean']>
  viewerIsFollowing?: Maybe<Scalars['Boolean']>
  /** @deprecated isFriend has been replaced with viewerIsFollowing */
  isFriend?: Maybe<Scalars['Boolean']>
  picture?: Maybe<Scalars['String']>
  profile: Profile
  sharedArticles: FeedArticle[]
  sharedArticlesCount?: Maybe<Scalars['Int']>
  sharedHighlightsCount?: Maybe<Scalars['Int']>
  sharedNotesCount?: Maybe<Scalars['Int']>
  friendsCount?: Maybe<Scalars['Int']>
  followersCount?: Maybe<Scalars['Int']>
}

export interface Profile {
  __typename?: 'Profile'
  id: Scalars['ID']
  username: Scalars['String']
  private: Scalars['Boolean']
  bio?: Maybe<Scalars['String']>
  pictureUrl?: Maybe<Scalars['String']>
}
// ...
export interface LinkShareInfo {
  __typename?: 'LinkShareInfo'
  title: Scalars['String']
  description: Scalars['String']
  imageUrl: Scalars['String']
}

export interface Link {
  __typename?: 'Link'
  id: Scalars['ID']
  url: Scalars['String']
  slug: Scalars['String']
  savedBy: User
  savedAt: Scalars['Date']
  updatedAt: Scalars['Date']
  savedByViewer: Scalars['Boolean']
  postedByViewer: Scalars['Boolean']
  readState: ReadState
  highlightStats: HighlightStats
  shareInfo: LinkShareInfo
  shareStats: ShareStats
  page: Page
}

export enum PageType {
  Article = 'ARTICLE',
  Book = 'BOOK',
  File = 'FILE',
  Profile = 'PROFILE',
  Website = 'WEBSITE',
  Highlights = 'HIGHLIGHTS',
  Unknown = 'UNKNOWN',
  Tweet = 'TWEET',
  Video = 'VIDEO',
  Image = 'IMAGE'
}
// ...
export interface RecommendingUser {
  __typename?: 'RecommendingUser'
  userId: Scalars['String']
  name: Scalars['String']
  username: Scalars['String']
  profileImageURL?: Maybe<Scalars['String']>
}

export interface Recommendation {
  __typename?: 'Recommendation'
  id: Scalars['ID']
  name: Scalars['String']
  user?: Maybe<RecommendingUser>
  recommendedAt: Scalars['Date']
  note?: Maybe<Scalars['String']>
}

export interface Article {
  __typename?: 'Article'
  id: Scalars['ID']
  title: Scalars['String']
  slug: Scalars['String']
  url: Scalars['String']
  hash: Scalars['String']
  content: Scalars['String']
  pageType?: Maybe<PageType>
  contentReader: ContentReader
  hasContent?: Maybe<Scalars['Boolean']>
  author?: Maybe<Scalars['String']>
  image?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  originalHtml?: Maybe<Scalars['String']>
  createdAt: Scalars['Date']
  savedAt: Scalars['Date']
  updatedAt: Scalars['Date']
  publishedAt?: Maybe<Scalars['Date']>
  readingProgressTopPercent?: Maybe<Scalars['Float']>
  readingProgressPercent: Scalars['Float']
  readingProgressAnchorIndex: Scalars['Int']
  sharedComment?: Maybe<Scalars['String']>
  savedByViewer?: Maybe<Scalars['Boolean']>
  postedByViewer?: Maybe<Scalars['Boolean']>
  originalArticleUrl?: Maybe<Scalars['String']>
  highlights: Highlight[]
  shareInfo?: Maybe<LinkShareInfo>
  isArchived: Scalars['Boolean']
  linkId?: Maybe<Scalars['ID']>
  labels?: Maybe<Label[]>
  uploadFileId?: Maybe<Scalars['ID']>
  siteName?: Maybe<Scalars['String']>
  siteIcon?: Maybe<Scalars['String']>
  subscription?: Maybe<Scalars['String']>
  unsubMailTo?: Maybe<Scalars['String']>
  unsubHttpUrl?: Maybe<Scalars['String']>
  state?: Maybe<ArticleSavingRequestStatus>
  language?: Maybe<Scalars['String']>
  readAt?: Maybe<Scalars['Date']>
  recommendations?: Maybe<Recommendation[]>
  wordsCount?: Maybe<Scalars['Int']>
}

export interface ArticleHighlightsArgs {
  input?: Maybe<ArticleHighlightsInput>
}

export type ArticleResult = ArticleSuccess | ArticleError

export enum ArticleErrorCode {
  NotFound = 'NOT_FOUND',
  BadData = 'BAD_DATA',
  Unauthorized = 'UNAUTHORIZED'
}

export interface ArticleError {
  __typename?: 'ArticleError'
  errorCodes: ArticleErrorCode[]
}

export interface ArticleSuccess {
  __typename?: 'ArticleSuccess'
  article: Article
}
// ...
export enum HighlightType {
  Highlight = 'HIGHLIGHT',
  Redaction = 'REDACTION',
  Note = 'NOTE'
}

export interface Highlight {
  __typename?: 'Highlight'
  id: Scalars['ID']
  shortId: Scalars['String']
  user: User
  quote?: Maybe<Scalars['String']>
  prefix?: Maybe<Scalars['String']>
  suffix?: Maybe<Scalars['String']>
  patch?: Maybe<Scalars['String']>
  annotation?: Maybe<Scalars['String']>
  replies: HighlightReply[]
  sharedAt?: Maybe<Scalars['Date']>
  createdAt: Scalars['Date']
  updatedAt: Scalars['Date']
  reactions: Reaction[]
  createdByMe: Scalars['Boolean']
  highlightPositionPercent?: Maybe<Scalars['Float']>
  highlightPositionAnchorIndex?: Maybe<Scalars['Int']>
  labels?: Maybe<Label[]>
  type: HighlightType
  html?: Maybe<Scalars['String']>
}
// ...
export interface HighlightReply {
  __typename?: 'HighlightReply'
  id: Scalars['ID']
  user: User
  highlight: Highlight
  text: Scalars['String']
  createdAt: Scalars['Date']
  updatedAt: Scalars['Date']
}
// ...
export interface Reaction {
  __typename?: 'Reaction'
  id: Scalars['ID']
  user: User
  code: ReactionType
  createdAt: Scalars['Date']
  updatedAt?: Maybe<Scalars['Date']>
}
// ...
export enum ArticleSavingRequestStatus {
  Processing = 'PROCESSING',
  Succeeded = 'SUCCEEDED',
  Failed = 'FAILED',
  Deleted = 'DELETED',
  Archived = 'ARCHIVED'
}
// ...
export interface Label {
  __typename?: 'Label'
  id: Scalars['ID']
  name: Scalars['String']
  color: Scalars['String']
  description?: Maybe<Scalars['String']>
  createdAt?: Maybe<Scalars['Date']>
  position?: Maybe<Scalars['Int']>
  internal?: Maybe<Scalars['Boolean']>
}
// ...
export interface SearchItem {
  __typename?: 'SearchItem'
  id: Scalars['ID']
  title: Scalars['String']
  slug: Scalars['String']
  url: Scalars['String']
  pageType: PageType
  contentReader: ContentReader
  createdAt: Scalars['Date']
  updatedAt?: Maybe<Scalars['Date']>
  isArchived: Scalars['Boolean']
  readingProgressTopPercent?: Maybe<Scalars['Float']>
  readingProgressPercent: Scalars['Float']
  readingProgressAnchorIndex: Scalars['Int']
  author?: Maybe<Scalars['String']>
  image?: Maybe<Scalars['String']>
  description?: Maybe<Scalars['String']>
  publishedAt?: Maybe<Scalars['Date']>
  ownedByViewer?: Maybe<Scalars['Boolean']>
  originalArticleUrl?: Maybe<Scalars['String']>
  uploadFileId?: Maybe<Scalars['ID']>
  pageId?: Maybe<Scalars['ID']>
  shortId?: Maybe<Scalars['String']>
  quote?: Maybe<Scalars['String']>
  annotation?: Maybe<Scalars['String']>
  labels?: Maybe<Label[]>
  subscription?: Maybe<Scalars['String']>
  unsubMailTo?: Maybe<Scalars['String']>
  unsubHttpUrl?: Maybe<Scalars['String']>
  state?: Maybe<ArticleSavingRequestStatus>
  siteName?: Maybe<Scalars['String']>
  language?: Maybe<Scalars['String']>
  readAt?: Maybe<Scalars['Date']>
  savedAt: Scalars['Date']
  highlights?: Maybe<Highlight[]>
  siteIcon?: Maybe<Scalars['String']>
  recommendations?: Maybe<Recommendation[]>
  wordsCount?: Maybe<Scalars['Int']>
  content?: Maybe<Scalars['String']>
  archivedAt?: Maybe<Scalars['Date']>
}

export interface SearchItemEdge {
  __typename?: 'SearchItemEdge'
  cursor: Scalars['String']
  node: SearchItem
}

export interface SearchSuccess {
  __typename?: 'SearchSuccess'
  edges: SearchItemEdge[]
  pageInfo: PageInfo
}